+++ app/src/main/java/reikai/presentation/details/EntryDetailsSkeleton.kt
@@ -29,7 +29,6 @@ import androidx.compose.ui.draw.alpha
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.semantics.contentDescription
 import androidx.compose.ui.semantics.semantics
-import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import tachiyomi.i18n.MR
 import tachiyomi.presentation.core.i18n.stringResource
@@ -87,10 +86,10 @@ fun EntryDetailsSkeleton(modifier: Modifier = Modifier) {
 }
 
 @Composable
-private fun Bone(modifier: Modifier, corner: Dp = 4.dp) {
+private fun Bone(modifier: Modifier) {
     Box(
         modifier
-            .clip(RoundedCornerShape(corner))
+            .clip(RoundedCornerShape(4.dp))
             .background(MaterialTheme.colorScheme.surfaceContainerHighest),
     )
 }